requirements.txt
requests==2.24.0
python-dotenv==0.14.0
pydash~=4.8.0
pymongo~=3.11.3
uvicorn~=0.11.8
pydantic~=1.6.1
starlette~=0.13.6
fastapi~=0.65.1
celery~=5.0.5
pendulum~=2.1.2
dnspython==2.1.0